# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 1.0.0
  description: Task Management Portal API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List users
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserOut"
    post:
      operationId: createUser
      tags: [users]
      summary: Create user
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserCreate"
      responses:
        "200":
          description: Created user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserOut"
  /users/{userId}:
    put:
      operationId: updateUser
      tags: [users]
      summary: Update user
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserOut"
  /students:
    get:
      operationId: listStudents
      tags: [students]
      summary: List students
      security:
        - bearerAuth: []
      parameters:
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of students
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StudentOut"
    post:
      operationId: createStudent
      tags: [students]
      summary: Create student
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StudentCreate"
      responses:
        "200":
          description: Created student
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudentOut"
  /students/{studentId}:
    put:
      operationId: updateStudent
      tags: [students]
      summary: Update student
      security:
        - bearerAuth: []
      parameters:
        - name: studentId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StudentUpdate"
      responses:
        "200":
          description: Updated student
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudentOut"
  /universities:
    get:
      operationId: listUniversities
      tags: [universities]
      summary: List universities
      security:
        - bearerAuth: []
      parameters:
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of universities
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UniversityOut"
    post:
      operationId: createUniversity
      tags: [universities]
      summary: Create university
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UniversityCreate"
      responses:
        "200":
          description: Created university
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UniversityOut"
  /universities/{uniId}:
    put:
      operationId: updateUniversity
      tags: [universities]
      summary: Update university
      security:
        - bearerAuth: []
      parameters:
        - name: uniId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UniversityUpdate"
      responses:
        "200":
          description: Updated university
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UniversityOut"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    UserOut:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        full_name:
          type: string
        role:
          type: string
        is_active:
          type: boolean
        created_at:
          type: string
          format: date-time
      required: [id, email, full_name, role, is_active, created_at]
    UserCreate:
      type: object
      properties:
        email:
          type: string
          format: email
        full_name:
          type: string
        password:
          type: string
        role:
          type: string
          default: agent
      required: [email, full_name, password]
    UserUpdate:
      type: object
      properties:
        email:
          type: string
        full_name:
          type: string
        role:
          type: string
        is_active:
          type: boolean
        password:
          type: string
    StudentOut:
      type: object
      properties:
        id:
          type: integer
        full_name:
          type: string
        passport_no:
          type: string
          nullable: true
        dob:
          type: string
          format: date
          nullable: true
        phone:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
      required: [id, full_name, created_at]
    StudentCreate:
      type: object
      properties:
        full_name:
          type: string
        passport_no:
          type: string
        dob:
          type: string
          format: date
        phone:
          type: string
        email:
          type: string
      required: [full_name]
    StudentUpdate:
      type: object
      properties:
        full_name:
          type: string
        passport_no:
          type: string
        dob:
          type: string
          format: date
        phone:
          type: string
        email:
          type: string
    UniversityOut:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        country:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
      required: [id, name, created_at]
    UniversityCreate:
      type: object
      properties:
        name:
          type: string
        country:
          type: string
      required: [name]
    UniversityUpdate:
      type: object
      properties:
        name:
          type: string
        country:
          type: string
...