# lib/api-zod/src/generated/types/studentCreate.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Task Management Portal API
 * OpenAPI spec version: 1.0.0
 */

export interface StudentCreate {
  full_name: string;
  passport_no?: string;
  dob?: Date;
  phone?: string;
  email?: string;
}