# lib/api-zod/src/generated/types/studentUpdate.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Task Management Portal API
 * OpenAPI spec version: 1.0.0
 */

export interface StudentUpdate {
  full_name?: string;
  passport_no?: string;
  dob?: Date;
  phone?: string;
  email?: string;
}