# lib/api-zod/src/generated/types/userUpdate.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Task Management Portal API
 * OpenAPI spec version: 1.0.0
 */

export interface UserUpdate {
  email?: string;
  full_name?: string;
  role?: string;
  is_active?: boolean;
  password?: string;
}